backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pymongo==4.8.0
networkx==3.3
python-multipart==0.0.9
google-genai==0.3.0